feat: Improvements for PR
Update component name, delete unused variables and change css
