+++ packages/x-components/src/components/index.ts
@@ -30,7 +30,7 @@ export { default as SnippetCallbacks } from './snippet-callbacks.vue';
 export { default as PageLoaderButton } from './page-loader-button.vue';
 export { default as PageSelector } from './page-selector.vue';
 export { default as DisplayClickProvider } from './display-click-provider.vue';
-export { default as CustomTeleport } from './custom-teleport.vue';
+export { default as BaseTeleport } from './base-teleport.vue';
 
 // Utils
 export * from './decorators/injection.consts';
